# openapi/tidal-player-api.yaml
openapi: 3.1.0
info:
  title: Littoral - Local API
  version: 0.1.0
  description: |
    API locale exposée par Littoral (lecteur Tidal Electron).

    - REST pour interroger l'état, contrôler la lecture, rechercher dans le catalogue, gérer la file d'attente.
    - WebSocket sur `/events` pour être notifié en temps réel des changements de playback (play, pause, changement de morceau, position, file d'attente, etc.).
servers:
  - url: http://127.0.0.1:7143
    description: Instance locale (port par défaut, configurable)
...
paths:
...
  /playback/volume:
    post:
      tags: [playback]
      operationId: setVolume
      summary: Régler le volume (0..100)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [volume]
              properties:
                volume:
                  type: integer
                  minimum: 0
                  maximum: 100
      responses:
        '204': { description: OK }
...
  /queue/enqueue:
    post:
      tags: [queue]
      operationId: enqueueTrack
      summary: Ajouter un morceau à la file d'attente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trackId]
              properties:
                trackId:
                  type: string
                  description: Identifiant Tidal du morceau
                position:
                  type: string
                  enum: [end, next]
                  default: end
      responses:
        '200':
          description: Élément ajouté
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Queue'
...
  /catalogue/tracks/{trackId}:
    get:
      tags: [catalogue]
      operationId: getTrack
      parameters:
        - in: path
          name: trackId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Track' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  responses:
    Unauthorized:
      description: Pas authentifié auprès de Tidal
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    Track:
      type: object
      required: [id, title, artists, durationSeconds]
      properties:
        id: { type: string }
        title: { type: string }
        artists:
          type: array
          items: { $ref: '#/components/schemas/Artist' }
        album: { $ref: '#/components/schemas/AlbumRef' }
        durationSeconds: { type: number }
        isrc: { type: string }
        explicit: { type: boolean }
        coverUrl: { type: string, format: uri }
        url: { type: string, format: uri, description: URL Tidal }
    Artist:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
        pictureUrl: { type: string, format: uri }
    AlbumRef:
      type: object
      required: [id, title]
      properties:
        id: { type: string }
        title: { type: string }
        coverUrl: { type: string, format: uri }
...
    QueueItem:
      type: object
      required: [id, trackId, addedAt]
      properties:
        id: { type: string, description: Identifiant local de l'élément de file }
        trackId: { type: string }
        addedAt: { type: string, format: date-time }
        track: { $ref: '#/components/schemas/Track' }
    Queue:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/QueueItem' }
...